Pushing for Future Reference
Been experimenting all day with no luck on getting these arrays of structures to function as they should. I'm gonna take the lazy way out for now because I am honestly not sure this is possible anymore. Pushing this now that way there is record of this code and I can always come back to it later